Rename Variant::OptionsPresentation to VariantOptionsPresenter
